Move the TyExpression and TyExpressionVariant definitions into the ty module.
