+++ sway-core/src/language/ty/expression/mod.rs
@@ -1,1 +1,6 @@
+#[allow(clippy::module_inception)]
+mod expression;
+mod expression_variant;
 
+pub use expression::*;
+pub use expression_variant::*;
