+++ sway-core/src/semantic_analysis/ast_node/expression/match_expression/typed/matcher.rs
@@ -1,13 +1,13 @@
 use crate::{
     error::{err, ok},
-    language::Literal,
+    language::{ty, Literal},
     semantic_analysis::{
         ast_node::expression::typed_expression::{
             instantiate_struct_field_access, instantiate_tuple_index_access,
             instantiate_unsafe_downcast,
         },
         namespace::Namespace,
-        IsConstant, TyEnumVariant, TyExpression, TyExpressionVariant, VariableMutability,
+        IsConstant, TyEnumVariant, VariableMutability,
     },
     type_system::unify,
     CompileResult, Ident, TypeId,
@@ -18,9 +18,9 @@ use sway_types::span::Span;
 use super::typed_scrutinee::{TyScrutinee, TyScrutineeVariant, TyStructScrutineeField};
 
 /// List of requirements that a desugared if expression must include in the conditional.
-pub(crate) type MatchReqMap = Vec<(TyExpression, TyExpression)>;
+pub(crate) type MatchReqMap = Vec<(ty::TyExpression, ty::TyExpression)>;
 /// List of variable declarations that must be placed inside of the body of the if expression.
-pub(crate) type MatchDeclMap = Vec<(Ident, TyExpression)>;
+pub(crate) type MatchDeclMap = Vec<(Ident, ty::TyExpression)>;
 /// This is the result type given back by the matcher.
 pub(crate) type MatcherResult = (MatchReqMap, MatchDeclMap);
 
@@ -65,7 +65,7 @@ pub(crate) type MatcherResult = (MatchReqMap, MatchDeclMap);
 /// ]
 /// ```
 pub(crate) fn matcher(
-    exp: &TyExpression,
+    exp: &ty::TyExpression,
     scrutinee: TyScrutinee,
     namespace: &mut Namespace,
 ) -> CompileResult<MatcherResult> {
@@ -98,14 +98,14 @@ pub(crate) fn matcher(
 }
 
 fn match_literal(
-    exp: &TyExpression,
+    exp: &ty::TyExpression,
     scrutinee: Literal,
     span: Span,
 ) -> CompileResult<MatcherResult> {
     let match_req_map = vec![(
         exp.to_owned(),
-        TyExpression {
-            expression: TyExpressionVariant::Literal(scrutinee),
+        ty::TyExpression {
+            expression: ty::TyExpressionVariant::Literal(scrutinee),
             return_type: exp.return_type,
             is_constant: IsConstant::No,
             span,
@@ -115,7 +115,7 @@ fn match_literal(
     ok((match_req_map, match_decl_map), vec![], vec![])
 }
 
-fn match_variable(exp: &TyExpression, scrutinee_name: Ident) -> CompileResult<MatcherResult> {
+fn match_variable(exp: &ty::TyExpression, scrutinee_name: Ident) -> CompileResult<MatcherResult> {
     let match_req_map = vec![];
     let match_decl_map = vec![(scrutinee_name, exp.to_owned())];
 
@@ -123,15 +123,15 @@ fn match_variable(exp: &TyExpression, scrutinee_name: Ident) -> CompileResult<Ma
 }
 
 fn match_constant(
-    exp: &TyExpression,
+    exp: &ty::TyExpression,
     scrutinee_name: Ident,
     scrutinee_type_id: TypeId,
     span: Span,
 ) -> CompileResult<MatcherResult> {
     let match_req_map = vec![(
         exp.to_owned(),
-        TyExpression {
-            expression: TyExpressionVariant::VariableExpression {
+        ty::TyExpression {
+            expression: ty::TyExpressionVariant::VariableExpression {
                 name: scrutinee_name,
                 span: span.clone(),
                 mutability: VariableMutability::Immutable,
@@ -147,7 +147,7 @@ fn match_constant(
 }
 
 fn match_struct(
-    exp: &TyExpression,
+    exp: &ty::TyExpression,
     fields: Vec<TyStructScrutineeField>,
     namespace: &mut Namespace,
 ) -> CompileResult<MatcherResult> {
@@ -190,7 +190,7 @@ fn match_struct(
 }
 
 fn match_enum(
-    exp: &TyExpression,
+    exp: &ty::TyExpression,
     variant: TyEnumVariant,
     scrutinee: TyScrutinee,
     span: Span,
@@ -210,7 +210,7 @@ fn match_enum(
 }
 
 fn match_tuple(
-    exp: &TyExpression,
+    exp: &ty::TyExpression,
     elems: Vec<TyScrutinee>,
     span: Span,
     namespace: &mut Namespace,
