+++ sway-core/src/semantic_analysis/ast_node/expression/match_expression/typed/typed_match_branch.rs
@@ -2,11 +2,11 @@ use sway_types::{Span, Spanned};
 
 use crate::{
     error::{err, ok},
-    language::parsed::MatchBranch,
+    language::{parsed::MatchBranch, ty},
     semantic_analysis::{
         ast_node::expression::match_expression::typed::typed_scrutinee::TyScrutinee, IsConstant,
-        TyAstNode, TyAstNodeContent, TyCodeBlock, TyExpression, TyExpressionVariant,
-        TyVariableDeclaration, TypeCheckContext, VariableMutability,
+        TyAstNode, TyAstNodeContent, TyCodeBlock, TyVariableDeclaration, TypeCheckContext,
+        VariableMutability,
     },
     type_system::insert_type,
     types::DeterministicallyAborts,
@@ -18,7 +18,7 @@ use super::matcher::{matcher, MatchReqMap};
 #[derive(Debug)]
 pub(crate) struct TyMatchBranch {
     pub(crate) conditions: MatchReqMap,
-    pub(crate) result: TyExpression,
+    pub(crate) result: ty::TyExpression,
     #[allow(dead_code)]
     span: Span,
 }
@@ -26,7 +26,7 @@ pub(crate) struct TyMatchBranch {
 impl TyMatchBranch {
     pub(crate) fn type_check(
         mut ctx: TypeCheckContext,
-        typed_value: &TyExpression,
+        typed_value: &ty::TyExpression,
         branch: MatchBranch,
     ) -> CompileResult<(TyMatchBranch, TyScrutinee)> {
         let mut warnings = vec![];
@@ -84,7 +84,7 @@ impl TyMatchBranch {
                 .by_ref()
                 .with_type_annotation(insert_type(TypeInfo::Unknown));
             check!(
-                TyExpression::type_check(ctx, result),
+                ty::TyExpression::type_check(ctx, result),
                 return err(warnings, errors),
                 warnings,
                 errors
@@ -104,19 +104,19 @@ impl TyMatchBranch {
         // of that code block to the block of code statements that we are already
         // generating. if the typed branch result is not a code block, then add
         // the typed branch result as an ast node to the block of code statements
-        let TyExpression {
+        let ty::TyExpression {
             expression: typed_result_expression_variant,
             return_type: typed_result_return_type,
             is_constant: typed_result_is_constant,
             span: typed_result_span,
         } = typed_result;
         match typed_result_expression_variant {
-            TyExpressionVariant::CodeBlock(TyCodeBlock { mut contents, .. }) => {
+            ty::TyExpressionVariant::CodeBlock(TyCodeBlock { mut contents, .. }) => {
                 code_block_contents.append(&mut contents);
             }
             typed_result_expression_variant => {
                 code_block_contents.push(TyAstNode {
-                    content: TyAstNodeContent::ImplicitReturnExpression(TyExpression {
+                    content: TyAstNodeContent::ImplicitReturnExpression(ty::TyExpression {
                         expression: typed_result_expression_variant,
                         return_type: typed_result_return_type,
                         is_constant: typed_result_is_constant,
@@ -129,8 +129,8 @@ impl TyMatchBranch {
 
         // assemble a new branch result that includes both the variable declarations
         // that we create and the typed result from the original untyped branch
-        let new_result = TyExpression {
-            expression: TyExpressionVariant::CodeBlock(TyCodeBlock {
+        let new_result = ty::TyExpression {
+            expression: ty::TyExpressionVariant::CodeBlock(TyCodeBlock {
                 contents: code_block_contents,
             }),
             return_type: typed_result.return_type,
