+++ sway-core/src/semantic_analysis/ast_node/expression/mod.rs
@@ -7,5 +7,4 @@ pub mod typed_expression_variant;
 pub use intrinsic_function::*;
 pub(crate) use match_expression::*;
 pub(crate) use struct_expr_field::*;
-pub(crate) use typed_expression::*;
 pub(crate) use typed_expression_variant::*;
