+++ sway-core/src/semantic_analysis/ast_node/expression/typed_expression/unsafe_downcast.rs
@@ -1,41 +1,37 @@
 use sway_types::Span;
 
 use crate::{
-    language::Literal,
+    language::{ty, Literal},
     semantic_analysis::{
         ast_node::expression::match_expression::MatchReqMap, IsConstant, TyEnumVariant,
-        TyExpressionVariant,
     },
     type_system::{insert_type, IntegerBits},
     TypeInfo,
 };
-
-use super::TyExpression;
-
 // currently the unsafe downcast expr is only used for enums, so this method is specialized for enums
 pub(crate) fn instantiate_unsafe_downcast(
-    exp: &TyExpression,
+    exp: &ty::TyExpression,
     variant: TyEnumVariant,
     span: Span,
-) -> (MatchReqMap, TyExpression) {
+) -> (MatchReqMap, ty::TyExpression) {
     let match_req_map = vec![(
-        TyExpression {
-            expression: TyExpressionVariant::EnumTag {
+        ty::TyExpression {
+            expression: ty::TyExpressionVariant::EnumTag {
                 exp: Box::new(exp.clone()),
             },
             return_type: insert_type(TypeInfo::UnsignedInteger(IntegerBits::SixtyFour)),
             is_constant: IsConstant::No,
             span: exp.span.clone(),
         },
-        TyExpression {
-            expression: TyExpressionVariant::Literal(Literal::U64(variant.tag as u64)),
+        ty::TyExpression {
+            expression: ty::TyExpressionVariant::Literal(Literal::U64(variant.tag as u64)),
             return_type: insert_type(TypeInfo::UnsignedInteger(IntegerBits::SixtyFour)),
             is_constant: IsConstant::No,
             span: exp.span.clone(),
         },
     )];
-    let unsafe_downcast = TyExpression {
-        expression: TyExpressionVariant::UnsafeDowncast {
+    let unsafe_downcast = ty::TyExpression {
+        expression: ty::TyExpressionVariant::UnsafeDowncast {
             exp: Box::new(exp.clone()),
             variant: variant.clone(),
         },
