+++ sway-core/src/semantic_analysis/storage_only_types.rs
@@ -1,17 +1,17 @@
+use sway_error::error::CompileError;
 use sway_types::{Span, Spanned};
 
-use crate::declaration_engine::declaration_engine::*;
-use crate::error::err;
-use crate::type_system::{is_type_info_storage_only, to_typeinfo, TypeId};
-use crate::{error::ok, semantic_analysis, CompileError, CompileResult, CompileWarning};
-use crate::{TyDeclaration, TyFunctionDeclaration, TypeInfo};
-
-use crate::semantic_analysis::{
-    TyAstNodeContent, TyConstantDeclaration, TyExpression, TyExpressionVariant,
-    TyIntrinsicFunctionKind, TyReassignment,
+use crate::{
+    declaration_engine::declaration_engine::*,
+    error::*,
+    language::ty,
+    semantic_analysis::{
+        TyAstNodeContent, TyCodeBlock, TyConstantDeclaration, TyDeclaration, TyEnumDeclaration,
+        TyFunctionDeclaration, TyImplTrait, TyIntrinsicFunctionKind, TyReassignment,
+        TyStorageDeclaration, TyStructDeclaration,
+    },
+    type_system::*,
 };
-
-use super::{TyEnumDeclaration, TyImplTrait, TyStorageDeclaration, TyStructDeclaration};
 
 fn ast_node_validate(x: &TyAstNodeContent) -> CompileResult<()> {
     let errors: Vec<CompileError> = vec![];
@@ -25,49 +25,49 @@ fn ast_node_validate(x: &TyAstNodeContent) -> CompileResult<()> {
     }
 }
 
-fn expr_validate(expr: &TyExpression) -> CompileResult<()> {
+fn expr_validate(expr: &ty::TyExpression) -> CompileResult<()> {
     let mut errors: Vec<CompileError> = vec![];
     let mut warnings: Vec<CompileWarning> = vec![];
     match &expr.expression {
-        TyExpressionVariant::Literal(_)
-        | TyExpressionVariant::VariableExpression { .. }
-        | TyExpressionVariant::FunctionParameter
-        | TyExpressionVariant::AsmExpression { .. }
-        | TyExpressionVariant::StorageAccess(_)
-        | TyExpressionVariant::AbiName(_) => (),
-        TyExpressionVariant::FunctionApplication { arguments, .. } => {
+        ty::TyExpressionVariant::Literal(_)
+        | ty::TyExpressionVariant::VariableExpression { .. }
+        | ty::TyExpressionVariant::FunctionParameter
+        | ty::TyExpressionVariant::AsmExpression { .. }
+        | ty::TyExpressionVariant::StorageAccess(_)
+        | ty::TyExpressionVariant::AbiName(_) => (),
+        ty::TyExpressionVariant::FunctionApplication { arguments, .. } => {
             for f in arguments {
                 check!(expr_validate(&f.1), continue, warnings, errors);
             }
         }
-        TyExpressionVariant::LazyOperator {
+        ty::TyExpressionVariant::LazyOperator {
             lhs: expr1,
             rhs: expr2,
             ..
         }
-        | TyExpressionVariant::ArrayIndex {
+        | ty::TyExpressionVariant::ArrayIndex {
             prefix: expr1,
             index: expr2,
         } => {
             check!(expr_validate(expr1), (), warnings, errors);
             check!(expr_validate(expr2), (), warnings, errors);
         }
-        TyExpressionVariant::IntrinsicFunction(TyIntrinsicFunctionKind {
+        ty::TyExpressionVariant::IntrinsicFunction(TyIntrinsicFunctionKind {
             arguments: exprvec,
             ..
         })
-        | TyExpressionVariant::Tuple { fields: exprvec }
-        | TyExpressionVariant::Array { contents: exprvec } => {
+        | ty::TyExpressionVariant::Tuple { fields: exprvec }
+        | ty::TyExpressionVariant::Array { contents: exprvec } => {
             for f in exprvec {
                 check!(expr_validate(f), continue, warnings, errors)
             }
         }
-        TyExpressionVariant::StructExpression { fields, .. } => {
+        ty::TyExpressionVariant::StructExpression { fields, .. } => {
             for f in fields {
                 check!(expr_validate(&f.value), continue, warnings, errors);
             }
         }
-        TyExpressionVariant::CodeBlock(cb) => {
+        ty::TyExpressionVariant::CodeBlock(cb) => {
             check!(
                 validate_decls_for_storage_only_types_in_codeblock(cb),
                 (),
@@ -75,7 +75,7 @@ fn expr_validate(expr: &TyExpression) -> CompileResult<()> {
                 errors
             );
         }
-        TyExpressionVariant::IfExp {
+        ty::TyExpressionVariant::IfExp {
             condition,
             then,
             r#else,
@@ -86,19 +86,19 @@ fn expr_validate(expr: &TyExpression) -> CompileResult<()> {
                 check!(expr_validate(r#else), (), warnings, errors);
             }
         }
-        TyExpressionVariant::StructFieldAccess { prefix: exp, .. }
-        | TyExpressionVariant::TupleElemAccess { prefix: exp, .. }
-        | TyExpressionVariant::AbiCast { address: exp, .. }
-        | TyExpressionVariant::EnumTag { exp }
-        | TyExpressionVariant::UnsafeDowncast { exp, .. } => {
+        ty::TyExpressionVariant::StructFieldAccess { prefix: exp, .. }
+        | ty::TyExpressionVariant::TupleElemAccess { prefix: exp, .. }
+        | ty::TyExpressionVariant::AbiCast { address: exp, .. }
+        | ty::TyExpressionVariant::EnumTag { exp }
+        | ty::TyExpressionVariant::UnsafeDowncast { exp, .. } => {
             check!(expr_validate(exp), (), warnings, errors)
         }
-        TyExpressionVariant::EnumInstantiation { contents, .. } => {
+        ty::TyExpressionVariant::EnumInstantiation { contents, .. } => {
             if let Some(f) = contents {
                 check!(expr_validate(f), (), warnings, errors);
             }
         }
-        TyExpressionVariant::WhileLoop { condition, body } => {
+        ty::TyExpressionVariant::WhileLoop { condition, body } => {
             check!(expr_validate(condition), (), warnings, errors);
             check!(
                 validate_decls_for_storage_only_types_in_codeblock(body),
@@ -107,9 +107,9 @@ fn expr_validate(expr: &TyExpression) -> CompileResult<()> {
                 errors
             );
         }
-        TyExpressionVariant::Break => (),
-        TyExpressionVariant::Continue => (),
-        TyExpressionVariant::Reassignment(reassignment) => {
+        ty::TyExpressionVariant::Break => (),
+        ty::TyExpressionVariant::Continue => (),
+        ty::TyExpressionVariant::Reassignment(reassignment) => {
             let TyReassignment {
                 lhs_base_name, rhs, ..
             } = &**reassignment;
@@ -121,7 +121,7 @@ fn expr_validate(expr: &TyExpression) -> CompileResult<()> {
             );
             check!(expr_validate(rhs), (), warnings, errors)
         }
-        TyExpressionVariant::StorageReassignment(storage_reassignment) => {
+        ty::TyExpressionVariant::StorageReassignment(storage_reassignment) => {
             let span = storage_reassignment.span();
             let rhs = &storage_reassignment.rhs;
             check!(
@@ -132,7 +132,7 @@ fn expr_validate(expr: &TyExpression) -> CompileResult<()> {
             );
             check!(expr_validate(rhs), (), warnings, errors)
         }
-        TyExpressionVariant::Return(stmt) => {
+        ty::TyExpressionVariant::Return(stmt) => {
             check!(expr_validate(&stmt.expr), (), warnings, errors)
         }
     }
@@ -299,9 +299,7 @@ pub fn validate_decls_for_storage_only_types_in_ast(ast_n: &TyAstNodeContent) ->
     ast_node_validate(ast_n)
 }
 
-pub fn validate_decls_for_storage_only_types_in_codeblock(
-    cb: &semantic_analysis::TyCodeBlock,
-) -> CompileResult<()> {
+pub fn validate_decls_for_storage_only_types_in_codeblock(cb: &TyCodeBlock) -> CompileResult<()> {
     let mut warnings: Vec<CompileWarning> = vec![];
     let mut errors: Vec<CompileError> = vec![];
     for x in &cb.contents {
