+++ sway-core/src/type_system/resolved_type.rs
@@ -1,6 +1,9 @@
-use crate::semantic_analysis::TyExpression;
-use crate::type_system::*;
-use crate::{language::CallPath, semantic_analysis::ast_node::TyStructField, Ident};
+use crate::{
+    language::{ty, CallPath},
+    semantic_analysis::ast_node::TyStructField,
+    type_system::*,
+    Ident,
+};
 use derivative::Derivative;
 
 #[derive(Derivative)]
@@ -32,7 +35,7 @@ pub enum ResolvedType {
     ContractCaller {
         abi_name: CallPath,
         #[derivative(PartialEq = "ignore", Hash = "ignore")]
-        address: Box<TyExpression>,
+        address: Box<ty::TyExpression>,
     },
     #[allow(dead_code)]
     Function {
