+++ sway-lsp/src/core/token.rs
@@ -1,14 +1,16 @@
 use dashmap::DashMap;
 use sway_core::{
-    language::parsed::{
-        Declaration, EnumVariant, Expression, FunctionDeclaration, FunctionParameter,
-        ReassignmentExpression, Scrutinee, StorageField, StructExpressionField, StructField,
-        TraitFn,
+    language::{
+        parsed::{
+            Declaration, EnumVariant, Expression, FunctionDeclaration, FunctionParameter,
+            ReassignmentExpression, Scrutinee, StorageField, StructExpressionField, StructField,
+            TraitFn,
+        },
+        ty,
     },
     semantic_analysis::ast_node::{
-        expression::typed_expression::TyExpression, TyDeclaration, TyEnumVariant,
-        TyFunctionDeclaration, TyFunctionParameter, TyReassignment, TyStorageField,
-        TyStorageReassignDescriptor, TyStructField, TyTraitFn,
+        TyDeclaration, TyEnumVariant, TyFunctionDeclaration, TyFunctionParameter, TyReassignment,
+        TyStorageField, TyStorageReassignDescriptor, TyStructField, TyTraitFn,
     },
     type_system::TypeId,
 };
@@ -59,7 +61,7 @@ pub enum AstToken {
 #[derive(Debug, Clone)]
 pub enum TypedAstToken {
     TypedDeclaration(TyDeclaration),
-    TypedExpression(TyExpression),
+    TypedExpression(ty::TyExpression),
     TypedFunctionDeclaration(TyFunctionDeclaration),
     TypedFunctionParameter(TyFunctionParameter),
     TypedStructField(TyStructField),
